+++ src/common/utils/constant.js
@@ -43,7 +43,6 @@ export const IEXEC_REQUEST_PARAMS = {
   IEXEC_RESULT_ENCRYPTION: 'iexec_result_encryption',
   IEXEC_RESULT_STORAGE_PROVIDER: 'iexec_result_storage_provider',
   IEXEC_RESULT_STORAGE_PROXY: 'iexec_result_storage_proxy',
-  IEXEC_DEVELOPER_LOGGER: 'iexec_developer_logger',
 };
 
 export const ANY = 'any';