+++ src/lib/IExecTaskModule.d.ts
@@ -36,7 +36,7 @@ interface Task {
 
 declare class TaskObservable extends Observable {
   /**
-   * subscribe to task updates via an Observer until either `complete()` or `error(error: Error)` is called on the Observer or the subscribtion is canceled by calling the returned unsubscribe method.
+   * subscribe to task updates via an Observer until either `complete()` or `error(error: Error)` is called on the Observer or the subscription is canceled by calling the returned unsubscribe method.
    *
    * return the `unsubscribe: () => void` method.
    *
@@ -165,6 +165,84 @@ export default class IExecTaskModule extends IExecModule {
    * ```
    */
   fetchResults(taskid: Taskid): Promise<Response>;
+  /**
+   * **SIGNER REQUIRED, ONLY REQUESTER**
+   *
+   * get the workers logs for specified task.
+   *
+   * _NB_: the workerpool must declare it's API url to enable this feature, check declared API url with `IExecWorkerpoolModule.getWorkerpoolApiUrl(workerpool)`
+   *
+   * example:
+   * ```js
+   * const logArray = await fetchLogs('0x668cb3e53ebbcc9999997709586c5af07f502f6120906fa3506ce1f531cedc81');
+   * logsArray.forEach(({ worker, stdout, stderr }) => {
+   *   console.log(`----- worker ${worker} -----`);
+   *   console.log(`stdout:\n${stdout}`);
+   *   console.log(`stderr:\n${stderr}`);
+   * });
+   * ```
+   */
+  fetchLogs(
+    taskid: Taskid,
+  ): Promise<Array<{ worker: Address; stdout: string; stderr: string }>>;
+  /**
+   * get off-chain status information for specified task.
+   *
+   * _NB_: the workerpool must declare it's API url to enable this feature, check declared API url with `IExecWorkerpoolModule.getWorkerpoolApiUrl(workerpool)`
+   *
+   * example:
+   * ```js
+   * const { task, replicates } = await fetchOffchainInfo('0x668cb3e53ebbcc9999997709586c5af07f502f6120906fa3506ce1f531cedc81');
+   *
+   * console.log(`task status: ${task.status}`);
+   * replicates.forEach(({ worker, status }) =>
+   *   console.log(`worker ${worker} replicate status: ${status}`)
+   * );
+   * ```
+   */
+  fetchOffchainInfo(taskid: Taskid): Promise<{
+    task: {
+      /**
+       * task status
+       *
+       * see https://protocol.docs.iex.ec/for-developers/task-feedback#task-statuses
+       */
+      status: string;
+      statusHistory: Array<{
+        /**
+         * task status
+         *
+         * see https://protocol.docs.iex.ec/for-developers/task-feedback#task-statuses
+         */
+        status: string;
+        date: string;
+        cause?: string;
+      }>;
+    };
+    replicates: Array<{
+      worker: Address;
+      /**
+       * app exit code (omitted when exit code is 0)
+       */
+      exitCode?: number;
+      /**
+       * replicate status
+       *
+       * see https://protocol.docs.iex.ec/for-developers/task-feedback#replicate-statuses
+       */
+      status: string;
+      statusHistory: Array<{
+        /**
+         * replicate status
+         *
+         * see https://protocol.docs.iex.ec/for-developers/task-feedback#replicate-statuses
+         */
+        status: string;
+        date: string;
+        cause?: string;
+      }>;
+    }>;
+  }>;
   /**
    * Create an IExecTaskModule instance using an IExecConfig instance
    */